backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
pandas==2.1.1
openai==0.28.1
requests==2.31.0
